(Chore): Redefined dir structure
